requirements.txt
boto3==1.34.137
Django==5.0.7
pandas==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.1.5
